xdg_shell: Move floating window with attach point
The buffer attach point was previously just used as an offset used
during buffer rendering, while the purpose of the attach point is to
move the window itself by the delta value.

Use the delta to move floating containers, ignore it for tiling
containers, and remove it as a variable in rendering.

Closes: https://github.com/swaywm/sway/issues/6020