idle_inhibit: Always listen for view_unmap
For wlr_idle_inhibit, we'd listen to inhibitor destroy, whereas for
manual idle inhibition, we'd listen for view unmap. If a view associated
with a wlr_idle_inhibit inhibitor was destroyed, we could possibly end
up with a dangling view pointer.

Always listen to view unmap to ensure we clean up properly.