Add keyboard shortcuts inhibitor ipc subscription
Add a new type of ipc subscription so clients can be notified of changes
in keyboard shortcuts inhibitors. This allows to react to those events
by e.g. changing container properties (title, ...) to notify the user
about this circumstance (add marker, change color, ...). For this reason
the event includes information about the affected container as well.

Signed-off-by: Michael Weiser <michael.weiser@gmx.de>