Avoid no-op output reconfiguring
When an output command doesn't actually change anything in the output
config the processing will still all be done and cause other changes.
Notably if you run a command like:

	output * dpms on

even if it does nothing it will still cause all the inputs to be
reconfigured and very noticeable amounts of input latency to occur.

To fix this save the current output config in the output data and on
configure, if the config is the same just return without doing anything
else.

Fixes #6280