input: Move get_current_time_msec in from util
get_current_time_msec is only used in cursor.c, so we can move it in and
make it static. This is primarily intended to avoid a symbol collision
with wlroots, which we unfortunately do not have a good solution for
yet.

(cherry picked from commit 80315217f76e923cb19cdb76aaf9a5d177f11f61)