Use fixed titlebar heights
Currently the height of titlebars can vary depending on their text.

If the text contains latin characters, then the height of those
characters (in whatever font applies) will define the titlebar height.
If a text contains CJK characters, the height of that font will
determine the height. The same applies for emoji and any other character
that's available only in the non-default font.

This issue is annoying in itself, but because all existing windows are
immediately resized, all applications need to rerender and this
frequently results in some of them crashing (Firefox, Chromium and
Signal are all curlpits here). I've only been able to reproduce this
with many windows open, so I suspect it's a matter of too much load for
the system to handle.

This quickly escalates to the point where visiting websites with emoji
in their title can kill all browser windows.

Additionally, when the titlebar of a window grows, it's first rendered
cropped, and the actual height only updates when there's a change in the
focused window (I'm sure other state changes trigger a redraw too), at
which point it's no longer cropped.

The approach taken here calculates the expected titlebar height based on
a short set of characters. Since these are likely to all use different
fonts, titlebars will be big enough to fit latin, CJK or emoji
characters without resizing.

I'm sure this still leaves out specific languages (e.g.: arabic
character bigger?). There's two ways to further improve this situation:

- Add a single character of each acceptable family to the same text
  used.
- Make the sample text user-configurable. This would allow a user to
  configure which font-families he expects to see, any anything bigger
  will be cropped. I don't feel the benefits are worth the effort here,
  but have nothing against it.