Prevent a layer shell infinite looping case
This commit try to fix a case we encounter when running swayout, a layer
shell to display text above the background.

When running wayout, other layer shell start to consume 100% of the CPU.
This seems related to a previous fix 744e85b. Then later, another commit
reworks this handle_surface_commit method 5fd5d643.

This is a tentative to fix the issue without missing the points of those
patches.

We basically try to not re-arrange layers if no layer changed.