input: Fix next/prev layout switch for grouped devices
Input devices in a seat can be grouped so that configuration changes on
one device, like the active layout, are automatically propagated to the
other devices.

The current logic for xkb_switch_layout wasn't accounting for this,
however. It was looping through all input devices and changing the
layout for each one, causing grouped devices to have their layouts
changed multiple times. This wasn't noticed when setting the layout to a
fixed index, but caused the wrong layout to be selected when using
the 'next' and 'prev' keywords.

In order to fix this issue, loop over each seat and, if the input
devices in the seat are grouped, change the layout only for the first
device in the seat.