output: update xcursor configuration on scale change
Sway will not correctly update the cursor glyph cache on an output
scale change. This results in sway falling back to a lower resolution
cursor glyph or just not rendering the correct cursor at all.

This is reproducible simply by using `swaymsg` to change
the scale sometimes, but I found 100% reproducibility when using
these patches[1].

[1] https://gitlab.freedesktop.org/wlroots/wlroots/-/merge_requests/3418