Fix ambiguity of 0 serial for surface configures
There is a bug where the transaction code will still wait for a client
commit even though the instruction configure is unsuccessful.
Unfortunately a serial of 0 is both used as a valid serial and
to denote an unsuccessful commit. Fix this my making relevant configure
functions return a boolean and taking in a serial pointer.