Revert "Fixed typo"
This reverts commit 26323d6ac63404274662f5da9e8e11c8052f7c81.

I was told this change shouldn't be done.

Also, I modified sway.5.scd according to the responses given by the developers.

N.B., This is my first time using git in this way. First I reverted the commit regarding a typo. Then I made changes to the file sway.5.scd. Lastly, I ran `git commit --amend -a`, and was presented with this file asking for a commit message. This is the same file I saw after reverting the old commit about a typo. I'm not sure if I did this correctly.