fix(autoname-workspace) Fix colon not getting removed after closing window
Fix a Bug resulting from faulty logic in the rename_workspaces where a whitespace
would have been added to the "icons" part of the workspace, no matters if
there are any icons present or not. This lead to the issue that not only a
trailing whitespace would've been erroneously added to the workspace's name,
but also that the colon separating the workspace number from the nonexistant
"icon" (only consisting of the whitespace) wouldn't disappear after closing
or moving all windows out of the workspace.