Fix memory leak of fallback_output and associated lists
This fixes memory leak of fallback_output because it was never freed
during the shutdown procedure.

This also fixes the memory leaks of lists associated with fallback_output.

This also removes one assertion condition because it was passed by
just setting a pointer reference to null but it left the heap content in place.