signal readiness to sd-daemon
This allows systemd to notice sway's startup is complete which allows,
for example, the session manager to start Wayland programs in the
right order. Without this signal, users have to go through rather
horrible hacks to tell systemd that it can start further units.

I've been using `NotifyAccess=all` in the sway.service file and `exec
systemd-notify --ready` in my sway config to emulate this, but it's
racy and error-prone.

A particularly nasty bug triggered by `NotifyAccess=all` in particular
is when podman starts and then terminates a container. In that
context, conmon(8) ends up notifying systemd it's the session master
and takes over thee "Main PID" field in systemd. When it dies, systemd
believes the session is over and proceeds to kill the entire
session. This is explained in more details in:

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1039857

This might not actually be the right place to do this. We call
`server_run` right after, and maybe there would be a better place. But
`server_run` only calls `wl_display_run` and that's part of the core
`wayland` library. I'm not sure Wayland itself is a place to do this,
so for now I'm scratching my own itch and doing this in Sway itself.